+++ .clean-publish.js
@@ -1,6 +1,0 @@
-'use strict';
-
-module.exports = {
-  files: ['.idea', 'docs', 'examples', 'flow-typed'],
-  packageManager: 'npm',
-};