+++ lint-staged.config.js
@@ -1,7 +1,0 @@
-'use strict';
-
-module.exports = {
-  '*.md': 'yaspeller-ci',
-  '**/*.js': ['eslint --fix', 'git add'],
-  '**/*.json': ['prettier --write', 'git add'],
-};