+++ prettier.config.js
@@ -1,8 +1,0 @@
-'use strict';
-
-module.exports = {
-  printWidth: 80,
-  tabWidth: 2,
-  singleQuote: true,
-  trailingComma: 'all',
-};