+++ src/utils.ts
@@ -1,0 +1,13 @@
+import {StyleProp} from 'react-native';
+
+const truncateString = (codeValue: string, codeLength: number) =>
+  codeValue.substr(0, codeLength);
+
+export const getSymbols = (codeValue: string, codeLength: number) =>
+  [
+    ...truncateString(codeValue, codeLength).split(''),
+    ...new Array(codeLength).fill(''),
+  ].slice(0, codeLength);
+
+export const getStyle = (base: StyleProp<any>, custom: StyleProp<any>) =>
+  custom ? [base, custom] : base;